+++ gui/views/rules_screen.py
@@ -1,33 +1,21 @@
-import os
-import time
-
 import arcade
 import arcade.gui
-
-from gui.screen_configuration import ScreenConfiguration
 import gui.view_manager
 
 
-class Rules(arcade.gui.UITextArea):
-    def __int__(self):
-        super(Rules, self)
-        self.fit_content()
+
+# class Rules(arcade.gui.UITextArea):
+#     def __int__(self):
+#         super(Rules, self)
+#         self.fit_content()
+from gui.screen_configuration import ScreenConfiguration
 
 
 class RulesView(arcade.View):
     def __init__(self, config: ScreenConfiguration):
         super().__init__()
+
         self.config = config
-
-        # get current working directory
-        cwd = os.getcwd()
-
-        # create a path to the file
-        path = os.path.join(cwd, "resources")
-
-        # open File and read Rules
-        with open(f'{path}/Rules.txt', 'r', encoding='UTF-8') as f:
-            self.__rules = f.read()
 
         # --- Required for all code that uses UI element,
         # a UIManager to handle the UI.
@@ -37,14 +25,21 @@ class RulesView(arcade.View):
         # Create Vertical Box to place the items in
         self.v_box = arcade.gui.UIBoxLayout()
 
+        # open File and read Rules
+        f = open('../../resources/Rules.txt', 'r', encoding='UTF-8')
+        self.rules = f.read()
+        f.close()
+
         arcade.set_background_color(arcade.color.WHITE_SMOKE)
 
         # Text Field to be put in V_Box
-        rules = Rules(self.config.width / 2, self.config.height / 2,
-                      self.config.width * 0.7, self.config.height * 0.7, self.__rules, 'arial', 25,
-                      arcade.color.BLACK, True, 5.5)
+        self.rules = arcade.gui.UITextArea(self.config.width / 2, self.config.height / 2,
+                      self.config.width * 0.7, self.config.height * 0.7, self.rules, 'arial', 25,
+                      arcade.color.BLACK)
+        self.rules.fit_content()
+        self.rules.scroll_speed = 5.5
 
-        self.v_box.add(rules)
+        self.v_box.add(self.rules)
 
         # Create a widget to hold the v_box widget, that will center the buttons
         self.manager.add(
